# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talent Bridge API",
    "version": "1.0.0",
    "description": "API documentation for the Talent Bridge platform",
    "contact": {
      "name": "Talent Bridge Team",
      "email": "support@talentbridge.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "Job": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64f789abc123def456789012"
          },
          "jobTitle": {
            "type": "string",
            "example": "Software Engineer"
          },
          "department": {
            "type": "string",
            "example": "Engineering"
          },
          "location": {
            "type": "string",
            "example": "Lahore"
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "Full-time",
              "Part-time",
              "Internship",
              "Contract"
            ],
            "example": "Full-time"
          },
          "salary": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number",
                "example": 50000
              },
              "max": {
                "type": "number",
                "example": 100000
              },
              "currency": {
                "type": "string",
                "example": "PKR"
              }
            }
          },
          "jobDescription": {
            "type": "string",
            "example": "Develop and maintain web applications."
          },
          "skillsRequired": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "skill": {
                  "type": "string",
                  "example": "JavaScript"
                },
                "proficiency": {
                  "type": "string",
                  "enum": [
                    "Beginner",
                    "Intermediate",
                    "Advanced"
                  ],
                  "example": "Intermediate"
                }
              }
            }
          },
          "benefits": {
            "type": "string",
            "example": "Health insurance, Flexible hours"
          },
          "category": {
            "type": "string",
            "example": "Technology"
          },
          "applicationDeadline": {
            "type": "string",
            "format": "date",
            "example": "2025-12-31"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "closed",
              "expired"
            ],
            "example": "active"
          },
          "companyId": {
            "type": "string",
            "description": "Employer company ID",
            "example": "64f890abc123def456789013"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Job creation timestamp",
            "example": "2025-08-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Job last update timestamp",
            "example": "2025-08-19T14:20:00.000Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Always false for error responses",
            "example": false
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP error status code",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Validation error or operation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Detailed error messages (optional)",
            "example": [
              "Email is required",
              "Password must be at least 6 characters"
            ]
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Current page limit",
            "example": 20
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "example": 150
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 8
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether there are more results available"
          },
          "nextCursor": {
            "type": "string",
            "description": "Cursor for the next page"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "summary": "Get all job posts with pagination and filtering",
        "description": "Retrieve a paginated list of active job posts with optional filtering by location, employment type, and category",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "description": "Number of jobs per page (max 50)",
            "example": 20
          },
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "description": "Filter by job location",
            "example": "Lahore"
          },
          {
            "in": "query",
            "name": "employmentType",
            "schema": {
              "type": "string",
              "enum": [
                "Full-time",
                "Part-time",
                "Internship",
                "Contract"
              ]
            },
            "description": "Filter by employment type",
            "example": "Full-time"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by job category",
            "example": "Technology"
          },
          {
            "in": "query",
            "name": "salaryMin",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "description": "Minimum salary filter",
            "example": 50000
          },
          {
            "in": "query",
            "name": "salaryMax",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "description": "Maximum salary filter",
            "example": 150000
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobs": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Job"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationInfo"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Jobs retrieved successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No jobs found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new job post",
        "description": "Create a new job posting (Only available to verified employers)",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jobTitle",
                  "department",
                  "location",
                  "employmentType",
                  "jobDescription",
                  "category",
                  "applicationDeadline"
                ],
                "properties": {
                  "jobTitle": {
                    "type": "string",
                    "description": "Job position title",
                    "example": "Senior Software Engineer"
                  },
                  "department": {
                    "type": "string",
                    "description": "Department or team",
                    "example": "Engineering"
                  },
                  "location": {
                    "type": "string",
                    "description": "Job location (city/remote)",
                    "example": "Lahore"
                  },
                  "employmentType": {
                    "type": "string",
                    "enum": [
                      "Full-time",
                      "Part-time",
                      "Internship",
                      "Contract"
                    ],
                    "description": "Type of employment",
                    "example": "Full-time"
                  },
                  "salary": {
                    "type": "object",
                    "description": "Salary range information",
                    "properties": {
                      "min": {
                        "type": "number",
                        "description": "Minimum salary",
                        "example": 80000
                      },
                      "max": {
                        "type": "number",
                        "description": "Maximum salary",
                        "example": 150000
                      },
                      "currency": {
                        "type": "string",
                        "description": "Currency code",
                        "example": "PKR"
                      }
                    }
                  },
                  "jobDescription": {
                    "type": "string",
                    "description": "Detailed job description and responsibilities",
                    "example": "We are looking for a Senior Software Engineer to join our team. You will be responsible for developing scalable web applications, mentoring junior developers, and contributing to architectural decisions."
                  },
                  "skillsRequired": {
                    "type": "array",
                    "description": "Required skills and proficiency levels",
                    "items": {
                      "type": "object",
                      "properties": {
                        "skill": {
                          "type": "string",
                          "example": "JavaScript"
                        },
                        "proficiency": {
                          "type": "string",
                          "enum": [
                            "Beginner",
                            "Intermediate",
                            "Advanced"
                          ],
                          "example": "Advanced"
                        }
                      }
                    },
                    "example": [
                      {
                        "skill": "JavaScript",
                        "proficiency": "Advanced"
                      },
                      {
                        "skill": "React",
                        "proficiency": "Intermediate"
                      },
                      {
                        "skill": "Node.js",
                        "proficiency": "Intermediate"
                      }
                    ]
                  },
                  "benefits": {
                    "type": "string",
                    "description": "Job benefits and perks",
                    "example": "Health insurance, Flexible working hours, Remote work options, Professional development budget"
                  },
                  "category": {
                    "type": "string",
                    "description": "Job category",
                    "example": "Technology"
                  },
                  "applicationDeadline": {
                    "type": "string",
                    "format": "date",
                    "description": "Application deadline (YYYY-MM-DD)",
                    "example": "2025-12-31"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "job": {
                          "$ref": "#/components/schemas/Job"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Job created successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only verified employers can create jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/search": {
      "get": {
        "summary": "Search jobs with advanced filtering",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search term for job title, description, or company"
          },
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "description": "Location filter"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Job category filter"
          },
          {
            "in": "query",
            "name": "employmentType",
            "schema": {
              "type": "string"
            },
            "description": "Employment type filter"
          },
          {
            "in": "query",
            "name": "minSalary",
            "schema": {
              "type": "number"
            },
            "description": "Minimum salary filter"
          },
          {
            "in": "query",
            "name": "maxSalary",
            "schema": {
              "type": "number"
            },
            "description": "Maximum salary filter"
          },
          {
            "in": "query",
            "name": "skills",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated skills filter"
          }
        ]
      }
    },
    "/api/v1/jobs/my/posts": {
      "get": {
        "summary": "Get jobs posted by current employer",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/jobs/{id}/status": {
      "patch": {
        "summary": "Update job status",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "closed",
                      "expired"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}